+++ .gitignore
@@ -18,7 +18,8 @@
 *.zip
 *.tar.gz
 *.rar
-
+*.txt
+*.properties
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
