+++ src/main/java/com/zs/basicprogram/Emi.java
@@ -1,5 +1,7 @@
-/*
-This class use for check the Emi
+package main.java.com.zs.basicprogram;
+
+/**
+ * This class help you to check the emi.
  */
 public class Emi {
     void emi(int loan_amount, int interest ,int time_period) {
@@ -8,6 +10,14 @@ public class Emi {
         interest = interest / (12 * 100);
         time_period = time_period * 12;
 
+        /**
+         * formula to calculate emi
+         *      E = (P.r.(1+r)n) / ((1+r)n – 1)
+         *      Here,
+         *      P = loan amount i.e principal amount
+         *      R = Interest rate per month
+         *      T = Loan time period in year
+         */
         emi = (loan_amount * interest * (float)Math.pow(1 + interest, time_period))
                 / (float)(Math.pow(1 + interest, time_period) - 1);
