+++ src/main/java/com/zs/basicprogram/ExtractSentence.java
@@ -1,21 +1,25 @@
-package com.zs.exercise1;
-
-/*
-This class is use for ectract the word in sentence
- */
+package main.java.com.zs.basicprogram;
 
 import java.util.ArrayList;
 
+/**
+ * This class help you to extract the words in given sentence.
+ */
 public class ExtractSentence {
-    void ExtractSentence(String sentence) {
+    void extractSentence(String sentence) {
         String st = "";
         int len = sentence.length();
 
-        // use array list to store the word
+        /**
+         * Use arraylist to store the words in sentence.
+         */
         ArrayList<String> words = new ArrayList<>();
 
         for(int i=0;i<len;i++) {
             if(sentence.charAt(i) == ' ') {
+                /**
+                 * Add the word in arraylist when space come because space divide the word in string.
+                 */
                 words.add(st);
                 st = "";
             }else {
