+++ src/main/java/com/zs/bodmas/Main.java
@@ -1,0 +1,15 @@
+package main.java.com.zs.bodmas;
+
+
+import java.util.Scanner;
+
+public class Main {
+    public static void main(String[] args) {
+	// write your code
+        Scanner in = new Scanner(System.in);
+        System.out.print("Enter the valid expression = ");
+        String exp = in.nextLine();
+
+        System.out.println();
+    }
+}
